+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - systemd-networkd overlay with IPv6 support
 - `wwctl overlay info` lists the variables used by an overlay template
 - Added support for setting mtu and mac address on bonds with NetworkManager #2059
+- Added `wwctl <node|profile> unset` #1954
 
 ### Changed